Split off over half of Functions into 2nd file Function.
