+++ codegen/castro/src/main/java/com/raelity/astrolog/castro/Pass2.java
@@ -23,7 +23,7 @@ import com.raelity.astrolog.castro.antlr.AstroParser.TermAddressOfContext;
 import com.raelity.astrolog.castro.mems.AstroMem.Var;
 import com.raelity.astrolog.castro.mems.Registers;
 import com.raelity.astrolog.castro.tables.Functions;
-import com.raelity.astrolog.castro.tables.Functions.Function;
+import com.raelity.astrolog.castro.tables.Function;
 
 import static com.raelity.astrolog.castro.Constants.constant;
 import static com.raelity.astrolog.castro.Util.lookup;
