+++ codegen/castro/src/main/java/com/raelity/astrolog/castro/tables/Functions.java
@@ -1,15 +1,9 @@
 /* Copyright © 2023 Ernie Rael. All rights reserved */
 
-// This file is generated by a script: scipts/extract_functions
-// from astrolog's express.cpp which is copyrighted by:
-// Walter D. Pullen (Astara@msn.com, http://www.astrolog.org/astrolog.htm)
-
 package com.raelity.astrolog.castro.tables;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -18,26 +12,24 @@ import java.util.Objects;
 
 import org.antlr.v4.runtime.Token;
 
-import com.raelity.astrolog.castro.Pass3;
 import com.raelity.astrolog.castro.antlr.AstroParser.ExprFuncContext;
-import com.raelity.astrolog.castro.antlr.AstroParser.Func_callContext;
-import com.raelity.astrolog.castro.mems.AstroMem;
-import com.raelity.astrolog.castro.mems.AstroMem.Var;
-import com.raelity.astrolog.castro.mems.Macros;
-import com.raelity.astrolog.castro.mems.Registers;
-import com.raelity.astrolog.castro.mems.Switches;
+import com.raelity.astrolog.castro.tables.Function.AstrologFunction;
+import com.raelity.astrolog.castro.tables.Function.FunctionUsage;
+import com.raelity.astrolog.castro.tables.Function.UnknownFunction;
+import com.raelity.astrolog.castro.tables.Function.UserFunction;
+import com.raelity.astrolog.castro.tables.Function.MacroFunction;
+import com.raelity.astrolog.castro.tables.Function.SwitchFunction;
 
 import static com.raelity.astrolog.castro.Error.*;
-import static com.raelity.astrolog.castro.Util.lookup;
-import static com.raelity.astrolog.castro.Util.macroSwitchFuncArgs;
 import static com.raelity.astrolog.castro.Util.reportError;
 
+/**
+ * Manage the various types of functions.
+ */
 public class Functions
 {
 public static final String FUNC_ID_SWITCH = "switch";
 public static final String FUNC_ID_MACRO = "macro";
-
-private record FunctionUsage(Token id, int narg) {}
 
 /**
  * isConst - False means this given call is not a constant, ignore all else.<br>
@@ -47,7 +39,7 @@ private record FunctionUsage(Token id, int narg) {}
 public static record FunctionConstValue(boolean isConst, int realVal, int displayVal){};
 public static FunctionConstValue NOT_CONST_VALUE = new FunctionConstValue(false, 0, 0);
 
-private static void reportFuncNargError(Object ctx_or_token,
+static void reportFuncNargError(Object ctx_or_token,
                                         String name, int expect, int result)
 {
     reportError(FUNC_NARG, ctx_or_token,
@@ -55,7 +47,7 @@ private static void reportFuncNargError(Object ctx_or_token,
                 name, expect, result);
 }
 
-private static void reportUserFuncNargError(Object ctx_or_token,
+static void reportUserFuncNargError(Object ctx_or_token,
                                             String name, int expect, int result)
 {
     reportError(ctx_or_token,
@@ -67,142 +59,6 @@ private static String lc(String s)
 {
     return s.toLowerCase(Locale.ROOT);
 }
-
-    /**
-     * The base class for any function; codegen and checking.
-     * Could be regular Astrolog function, special func like switch/macro,
-     * castro function. User defined function (really a macro).
-     */
-    abstract public static class Function
-    {
-    protected final String funcName;
-    protected final int narg;
-
-    public Function(String funcName, int narg)
-    { this.funcName = funcName; this.narg = narg; }
-
-    public final String name() { return funcName; }
-    public final int narg() { return narg; }
-
-    /** Invalid means that this Function definition itself has a problem;
-     * it can't be used.
-     */
-    public boolean isInvalid() {
-        return false;
-    }
-
-    public boolean isBuiltin() {
-        return true;
-    };
-
-    public boolean isUnknown() {
-        return false;
-    }
-
-    List<FunctionUsage> getReferences() {
-        return Collections.emptyList();
-    }
-
-    public void addReference(Token id, int narg) {
-        // typically don't care about references.
-    }
-
-    /** Check for special func args; returns true if no further
-     * checking needed, doesn't mean there's not an error.
-     * Typically meaningfull where unusual arguments
-     * are expected, like switch/macro name. A false return doesn't
-     * mean there's a problem, only that further checking is expected.
-     */
-    public boolean isDoneReportSpecialFuncArgs(Func_callContext ctx)
-    {
-        return false;
-    }
-
-     /* @return null for variable, else macro/switch memSpace */
-    public AstroMem targetMemSpace() {
-        return null;
-    }
-
-    /**
-     * Check for correct number of args.
-     * This impl assumes expr args; see override in StringFunction,
-     * for string args.
-     * @return true if nargs is OK, else false if bad num args.
-     */
-    public boolean checkReportArgs(Func_callContext ctx)
-    {
-        if(ctx.args.size() != narg()) {
-            reportFuncNargError(ctx, ctx.id.getText(), narg(), ctx.args.size());
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * This returns a constant result of a given function invocation.
-     * The default is not constant, this may be overriden for functions
-     * that are constant given context.
-     * <p>
-     * NOTE that if being a constant is dependent on
-     * an allocated addresses, then this must return false before isAllocFrozen()
-     * is true.
-     */
-    public FunctionConstValue constValue(ExprFuncContext ctx) {
-        return NOT_CONST_VALUE;
-    }
-
-    /** generate code for this function call */
-    public abstract StringBuilder genFuncCall(
-            StringBuilder sb, ExprFuncContext ctx, List<String> args);
-
-    @Override
-    public String toString()
-    {
-        return "Function{name:" + name() + '}';
-    }
-
-    @Override
-    public int hashCode()
-    {
-        int hash = 3;
-        hash = 79 * hash + Objects.hashCode(this.funcName);
-        return hash;
-    }
-
-    @Override
-    public boolean equals(Object obj)
-    {
-        if(this == obj)
-            return true;
-        if(obj == null)
-            return false;
-        if(getClass() != obj.getClass())
-            return false;
-        final Function other = (Function)obj;
-        return Objects.equals(this.funcName, other.funcName);
-    }
-
-    } /////////// class Function
-
-
-    abstract public static class StringArgsFunction extends Function
-    {
-    public StringArgsFunction(String funcName, int narg)
-    {
-        super(funcName, narg);
-    }
-
-    @Override
-    public boolean checkReportArgs(Func_callContext ctx)
-    {
-        if(ctx.strs.size() != narg()) {
-            reportFuncNargError(ctx, ctx.id.getText(), narg(), ctx.strs.size());
-            return false;
-        }
-        return true;
-    }
-
-    } /////////// class StringArgsFunction
 
 ///////////////////////////////////////////////////////////////////////////////
 //
@@ -268,8 +124,21 @@ public static void addUnkownFunction(Token id, int narg)
         throw new IllegalArgumentException();
 }
 
-/** dummyFunction is a singleton DummyFunction */
+/**
+ * dummyFunction is a singleton DummyFunction. Used to avoid null return.
+ */
 private static Function dummyFunction = new DummyFunction();
+    private static class DummyFunction extends Function
+    {
+    public DummyFunction()
+    { super("#DummyFunction#", 0); }
+
+    @Override
+    public StringBuilder genFuncCall(StringBuilder sb, ExprFuncContext ctx,
+                                     List<String> args)
+        { sb.append("#DummyFunctionCall#");  return sb;}
+    @Override public boolean isInvalid() { return true; }
+    } /////////// DummyFunction
 
 /**
  * @return Function for the name
@@ -347,233 +216,5 @@ void add(String funcName, int narg, String types)
     Objects.nonNull(types);
     funcsModifiableMap.put(lc(funcName), new AstrologFunction(funcName, narg));
 }
-
-    private static class DummyFunction extends Function
-    {
-    public DummyFunction()
-    { super("#DummyFunction#", 0); }
-
-    @Override
-    public StringBuilder genFuncCall(StringBuilder sb, ExprFuncContext ctx,
-                                     List<String> args)
-        { sb.append("#DummyFunctionCall#");  return sb;}
-    @Override public boolean isInvalid() { return true; }
-    } /////////// DummyFunction
-
-    /** Handles almost all Astrolog builtin functions */
-    static class AstrologFunction extends Function
-    {
-
-    AstrologFunction(String funcName, int narg)
-    {
-        super(funcName, narg);
-    }
-
-    @Override
-    public AstroMem targetMemSpace()
-    {
-        return null;
-    }
-
-    @Override
-    public StringBuilder genFuncCall(
-            StringBuilder sb, ExprFuncContext ctx, List<String> args)
-    {
-        if(Ops.isAnyOp(funcName))
-            reportError(FUNC_CASTRO, ctx.fc.id,
-                        "using castro operator '%s' as a function",
-                        ctx.fc.id.getText());
-        sb.append(funcName).append(' ');
-
-        for(String arg : args) {
-            sb.append(arg);
-        }
-        return sb;
-    }
-    } /////////// AstrologFunction
-
-    /* ************************************************************* */
-
-    /**
-     * Record name and number of args.
-     * May be replaced by a user function.
-     * At the end of pass1 these removed if not converted.
-     */
-    private static class UnknownFunction extends Function
-    {
-    private List<FunctionUsage> references = new ArrayList<>();
-    public UnknownFunction(Token id, int narg)
-    {
-        super(id.getText(), narg);
-        references.add(new FunctionUsage(id, narg));
-    }
-
-    @Override
-    public boolean isUnknown() {
-        return true;
-    }
-
-    @Override
-    public boolean isBuiltin() {
-        return false;
-    };
-
-    @Override
-    List<FunctionUsage> getReferences() {
-        return Collections.unmodifiableList(references);
-    }
-
-    @Override
-    public void addReference(Token id, int narg) {
-        references.add(new FunctionUsage(id, narg));
-    }
-
-    /**
-     * Check for correct number of args.
-     * UnknownFunction not handled here; just say it's ok.
-     */
-    @Override
-    public boolean checkReportArgs(Func_callContext ctx)
-    {
-        return true;
-    }
-
-    @Override
-    public StringBuilder genFuncCall(StringBuilder sb, ExprFuncContext ctx,
-                                                           List<String> args)
-    {
-        throw new IllegalStateException();
-    }
-    }
-
-    /* ************************************************************* */
-
-    /**
-     * A UserFunction is basically a macro.
-     * The macro name must be followed by "( ... )", zero or more arguments;
-     * otherwise it is a simple macro.
-     */
-    private static class UserFunction extends Function
-    {
-    /** The declared argument names are the variables for the user func. */
-    // A sorted set, argument declaration order, would be nice.
-    private final List<String> argNames;
-    
-    private UserFunction(String funcName, List<String> argNames)
-    {
-        super(funcName, argNames.size());
-
-        if(new HashSet<>(argNames).size() != argNames.size())
-            throw new IllegalArgumentException();
-
-        this.argNames = List.copyOf(argNames);
-    }
-    
-    @Override
-    public boolean isBuiltin()
-    {
-        return false;
-    }
-
-    /**
-     * Check for correct number of args.
-     * A user function must match narg; no warning.
-     * @return true if nargs is OK, else false if bad num args.
-     */
-    @Override
-    public boolean checkReportArgs(Func_callContext ctx)
-    {
-        if(ctx.args.size() != narg()) {
-
-            reportUserFuncNargError(ctx, ctx.id.getText(), narg(), ctx.args.size());
-            return false;
-        }
-        return true;
-    }
-    
-    /** Assign the arguments to the macro's variables, invoke the macro.
-     * Create a list that has the code to initialize macro function arguments;
-     * for "m1(a + 13, b + 7)" it looks something like:
-     * <br> "= 123 Add @a 13"
-     * <br> "= 127 Add @b 7"
-     * <br> "Macro 1"
-     * <p>
-     * Then encapsulate the code into a single Astrolog "DO*" expression.
-     */
-    @Override
-    public StringBuilder genFuncCall(StringBuilder sb, ExprFuncContext ctx,
-                                                           List<String> args)
-    {
-        if(argNames.size() != args.size()) {
-
-            throw new IllegalArgumentException();
-        }
-
-        List<String> callMacro = new ArrayList<>();
-        Registers regs = lookup(Registers.class);
-        // Set up the macro parameters.
-        for(int i = 0; i < args.size(); i++) {
-            Var var = regs.getVar(argNames.get(i));
-            callMacro.add("= " + var.getAddr() + " " + args.get(i));
-        }
-        // Add the call to invoke the macro.
-        Var macro = lookup(Macros.class).getVar(name());
-        callMacro.add("Macro " + macro.getAddr() + " ");
-        Pass3.appendDo(sb, callMacro, () -> "UserFunctionCall");
-
-        return sb;
-    }
-
-    @Override
-    public String toString()
-    {
-        return String.format("UserFunction{name:%s, args:%s}",
-                             name(), argNames);
-    }
-    
-    }
-
-    /* ************************************************************* */
-    private static class SwitchFunction extends SwitchMacroFunction
-    {
-    public SwitchFunction() { super("Switch"); }
-
-    @Override public AstroMem targetMemSpace() { return lookup(Switches.class); }
-    } /////////// SwitchFunction
-
-    private static class MacroFunction extends SwitchMacroFunction
-    {
-    public MacroFunction() { super("Macro"); }
-
-    @Override public AstroMem targetMemSpace() { return lookup(Macros.class); }
-    } /////////// MacroFunction
-
-    /* ************************************************************* */
-    private static abstract class SwitchMacroFunction extends Function
-    {
-
-    public SwitchMacroFunction(String name)
-    {
-        super(name, 1);
-    }
-
-    @Override
-    public boolean isDoneReportSpecialFuncArgs(Func_callContext ctx)
-    {
-        return macroSwitchFuncArgs(ctx, targetMemSpace());
-    }
-
-    @Override
-    public StringBuilder genFuncCall(StringBuilder sb, ExprFuncContext ctx,
-                                     List<String> args)
-    {
-        if(args.size() == 1)
-            sb.append(name()).append(' ').append(args.get(0));
-        else
-            sb.append("#").append(name()).append(":args# ");
-        return sb;
-    }
-
-    } /////////// SwitchMacroFunction
 
 }
